👷(build) update project identifiers, dependencies
