+++ build.gradle.kts
@@ -3,57 +3,34 @@ import org.apache.tools.ant.filters.ReplaceTokens
 plugins {
     id("java")
     id("java-library")
-    id("com.github.johnrengelman.shadow") version("8.1.1")
+    id("com.github.johnrengelman.shadow") version ("8.1.1")
 }
 
-group = "dev.kscott"
+group = "live.webcrawls.profiles"
 version = "1.0-SNAPSHOT"
 
-configure<JavaPluginConvention> {
-    sourceCompatibility = JavaVersion.VERSION_17
-    targetCompatibility = sourceCompatibility
+
+java {
+    toolchain.languageVersion.set(JavaLanguageVersion.of(17))
 }
 
 repositories {
     mavenCentral()
-    maven {
-        name = "papermc-repo"
-        url = uri("https://papermc.io/repo/repository/maven-public/")
-    }
-    maven {
-        name = "sonatype"
-        url = uri("https://oss.sonatype.org/content/groups/public/")
-    }
-    maven {
-        name = "papi"
-        url = uri("https://repo.extendedclip.com/content/repositories/placeholderapi/")
-    }
+    maven("https://repo.papermc.io/repository/maven-public/")
+    maven("https://repo.extendedclip.com/content/repositories/placeholderapi/")
 }
 
 dependencies {
-    compileOnlyApi("org.checkerframework:checker-qual:3.5.0")
+    // DX
+    compileOnly("org.checkerframework:checker-qual:3.36.0")
+    compileOnly("com.google.guava:guava:31.1-jre")
 
-    compileOnlyApi("com.google.guava:guava:21.0")
-
-    compileOnly("com.destroystokyo.paper:paper-api:1.16.4-R0.1-SNAPSHOT")
-
-    compileOnly("me.clip:placeholderapi:2.10.9")
-
-    api("com.google.inject:guice:4.2.3")
-
-    implementation("org.jdbi:jdbi3-core:3.17.0")
-    api("com.github.ben-manes.caffeine:caffeine:2.8.6")
-
-    implementation("com.zaxxer:HikariCP:3.4.5")
-
-    implementation("org.spongepowered:configurate-hocon:4.0.0")
-
-    api("net.kyori:adventure-platform-bukkit:4.0.0-SNAPSHOT")
-    implementation("net.kyori:adventure-text-minimessage:4.1.0-SNAPSHOT") {
-        isTransitive = true
-    }
-
-    api("cloud.commandframework:cloud-paper:1.2.0")
+    // Minecraft
+    compileOnly("io.papermc.paper:paper-api:1.20.1-R0.1-SNAPSHOT")
+    compileOnly("me.clip:placeholderapi:2.11.3")
+    implementation("net.kyori:adventure-text-minimessage:4.14.0")
+    implementation("cloud.commandframework:cloud-paper:1.2.0")
+    implementation("org.spongepowered:configurate-hocon:4.1.2")
 }
 
 tasks {
@@ -77,7 +54,6 @@ tasks {
         }
 
         relocates(
-                "com.github.benmanes.caffeine",
                 "com.typesafe.config",
                 "com.zaxxer.hikari",
                 "com.google.inject",
