+++ src/main/java/live/webcrawls/profiles/ProfilesPlugin.java
@@ -1,0 +1,17 @@
+package live.webcrawls.profiles;
+
+import org.bukkit.plugin.java.JavaPlugin;
+
+/**
+ * The main PlayerProfiles plugin class
+ */
+public final class ProfilesPlugin extends JavaPlugin {
+
+    @Override
+    public void onEnable() {
+    }
+
+    @Override
+    public void onDisable() {
+    }
+}
